requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
gunicorn==20.1.0
idna==2.10
lxml==4.6.3
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.17.0
rae==0.0.9
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
tweepy==3.10.0
twython==3.8.2
urllib3==1.26.4
